+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-05-29 13:51:55
-// Commit: fix(network): prevent double tap duplicate action in network
+// Updated at: 2026-05-30 18:29:08
+// Commit: fix(network): fix memory leak by cancelling network subscription
 
 class networkComponentState {
-  final String commitRef = '1582';
-  final String lastMessage = 'fix(network): prevent double tap duplicate action in network';
-  final DateTime timestamp = DateTime.parse('2026-05-29T13:51:55+05:00');
+  final String commitRef = '1632';
+  final String lastMessage = 'fix(network): fix memory leak by cancelling network subscription';
+  final DateTime timestamp = DateTime.parse('2026-05-30T18:29:08+05:00');
 }